fix(designer): address review on custom dividers
- doSplit split every selected compartment in one normalized pass; the
  old per-compartment loop used stale ids (splitCompartment renumbers each
  call), so multi-compartment splits could miss compartments.
- Route to the authored-divider builder only when customGrid is present,
  matching slotBuilder (avoids empty pieces + parametric slots on a custom
  config missing its grid).
- Editor and the numbered 3D overlay now map the grid over the overhang-
  expanded interior (moved overhang math to shared/ so both sides share it),
  so custom previews are correct on overhang bins.
- Offset the snappable score fully outboard of the notch (slotWidth/2 +
  SNAP_SCORE_WIDTH/2) so the retained piece keeps a complete notch.
